Abstract to separate extract file
