+++ lib/trace.js
@@ -1,99 +1,11 @@
 "use strict";
 
 const { parse } = require("acorn-node");
-const walk = require("acorn-node/walk");
+const { getDeps } = require("./extract");
 const { exists, readFile } = require("./util");
 
-// TODO: Support
-// - require
-// - require.resolve
-// - static import
-
-const parseDeps = (ast) => {
-  const deps = new Set();
-
-  walk.simple(ast, {
-    // Node
-    CallExpression(node) {
-      // Populate callee with empty defaults.
-      const callee = {
-        object: {},
-        property: {},
-        ...node.callee
-      };
-
-      // Only get first argument if only one to ensure we have a single string
-      // literal.
-      const args = node.arguments || [];
-      const firstArg = args.length === 1 ? args[0] : {};
-
-      // Node: `require`
-      //
-      // ```js
-      // require("foo");
-      // ```
-      //
-      // ```yml
-      // type: CallExpression
-      // callee:
-      //   type: Identifier
-      //   name: require
-      // arguments:
-      // - type: Literal
-      //   value: foo
-      //   raw: '"foo"'
-      // ```
-      if (
-        callee.type === "Identifier" && callee.name === "require"
-        && firstArg.type === "Literal"
-      ) {
-        deps.add(firstArg.value);
-      }
-
-      // Node: `require.resolve`
-      //
-      // ```js
-      // require.resolve("foo");
-      // ```
-      //
-      // ```yml
-      // type: CallExpression
-      // callee:
-      //   type: MemberExpression
-      //   object:
-      //     type: Identifier
-      //     name: require
-      //   property:
-      //     type: Identifier
-      //     name: resolve
-      //   computed: false
-      // arguments:
-      // - type: Literal
-      //   value: foo
-      //   raw: '"foo"'
-      // ```
-      if (callee.type === "MemberExpression"
-        && callee.object.type === "Identifier" && callee.object.name === "require"
-        && callee.property.type === "Identifier" && callee.property.name === "resolve"
-        && firstArg.type === "Literal") {
-        deps.add(firstArg.value);
-      }
-    },
-
-    // ESM: `import`
-    ImportDeclaration({ source = {} }) {
-      if (source.type === "Literal") {
-        deps.add(source.value);
-      }
-    }
-
-    // ESM: `import()` TODO
-  });
-
-  return deps;
-};
-
 // TODO: Delete
+// eslint-disable-next-line no-unused-vars
 const SAMPLE_SRC_FILE = `
 import * as React from "react-esm";
 import "bar-esm";
@@ -129,10 +41,11 @@ const traceFile = async ({ file } = {}) => {
   }
 
   const src = await readFile(file);
-  const deps = parseDeps(parse(src, { sourceType: "module" }));
+  const deps = getDeps(parse(src, { sourceType: "module" }));
 
   // TODO: Resolve each dep to file location.
   // TODO: Recurse and read.
+  // TODO: Consider output format: List? Set? Tree?
 
   return Array.from(deps).sort();
 };
